Spring Security

+++ journal/src/main/java/com/bytesize/journal/service/journalEntryService.java
@@ -3,12 +3,14 @@ package com.bytesize.journal.service;
 import com.bytesize.journal.entity.User;
 import com.bytesize.journal.entity.journalEntry;
 import com.bytesize.journal.repository.journalEntryRepo;
+import com.bytesize.journal.serviceImpl.UserAuthServiceImpl;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,13 +23,17 @@ public class journalEntryService {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserAuthServiceImpl userAuthService;
+
     @Transactional
-    public void saveJournalEntry(journalEntry jEntry, String userName)
+    public void saveJournalEntry(journalEntry jEntry)
     {
-        User user = userService.findByUserName(userName);
+        User user = userService.findByUserName(userAuthService.getUserName());
         jEntry.setDate(LocalDateTime.now());
         journalEntry savedEntry = journalEntryRepo.save(jEntry);
         user.getJournalEntries().add(savedEntry);
+        userService.saveUser(user);
     }
 
     public List<journalEntry> getAll(){
@@ -39,9 +45,27 @@ public class journalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void deleteById(ObjectId id)
+    public boolean deleteById(ObjectId id)
     {
-        journalEntryRepo.deleteById(id);
+        User user = userService.findByUserName(userAuthService.getUserName());
+        boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveUser(user);
+            journalEntryRepo.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean checkUserHasJournalEntry(ObjectId id){
+        User user = userService.findByUserName(userAuthService.getUserName());
+        List<journalEntry> userLinkedJournalEntries = new ArrayList<>();
+        for (journalEntry x : user.getJournalEntries()) {
+            if (x.getId().equals(id)) {
+                userLinkedJournalEntries.add(x);
+            }
+        }
+        return userLinkedJournalEntries.isEmpty();
     }
 
 }
